# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: HoloNet API
  version: "1"
  description: |
    SNMP trap → messaging bridge. Every UI action maps to an endpoint here.

    ## Authentication
    Session cookie (`holonet_session`) issued by `/api/v1/auth/login` or
    `/api/v1/auth/setup`. When `auth.enabled` is `false` (e.g. behind Cloudflare
    Access) the API is open. First run: `/api/v1/auth/setup` creates the admin.

    Sealed secrets (community strings, channel credentials) are write-only and
    never returned in responses.
servers:
  - url: /
paths:
...
  /api/v1/auth/setup:
    post:
      summary: First-run admin setup
      security: []
      responses: { "201": { description: Created }, "409": { description: Already configured } }
  /api/v1/auth/login:
    post:
      summary: Log in
      security: []
      responses: { "200": { description: OK }, "401": { description: Invalid } }
...
  /api/v1/traps:
    get:
      summary: List traps (Events tab)
      parameters:
        - { name: sort, in: query, schema: { type: string, enum: [received_at, source_ip, resolved_name, severity, matched_rule, status, id] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: limit, in: query, schema: { type: integer } }
      responses: { "200": { description: OK } }
  /api/v1/traps/{id}:
    get: { summary: Get one trap, responses: { "200": { description: OK }, "404": { description: Not found } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: holonet_session
security:
  - cookieAuth: []